Split OpenApiDocumentTestsBase into feature-scoped partials
The single 780-line file mixed strings, numerics, collections,
composition, annotations, and components-cleanup tests behind ad-hoc
section banners. Split into one partial per feature so each concern
opens to roughly a screen of related tests, with nullable variants
sitting next to their non-nullable siblings instead of in a separate
nullables file. The base file keeps the abstract members, broken
flags, GetDocumentAsync, and Document_IsServed.
